+++ tests/unit/tensorflow/test_commands_tf2.py
@@ -158,6 +158,7 @@ def test_tf2_convert_tf_trt():
             workspace=Workspace(workspace),
             verbose=True,
             dataloader_trt_profile=TensorRTProfile(),
+            custom_args={},
         )
 
         assert command_output.status == CommandStatus.OK